+++ app/src/main/java/woowacourse/movie/data/db/ReservationDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data
+package woowacourse.movie.data.db
 
 import androidx.room.Dao
 import androidx.room.Insert
