+++ app/src/main/java/woowacourse/movie/presentation/home/reservation/seat/ReservationSeatFragment.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import android.view.View
 import androidx.core.os.bundleOf
 import woowacourse.movie.R
-import woowacourse.movie.data.ReservationDaoListenerImpl
-import woowacourse.movie.data.ReservationDatabase
+import woowacourse.movie.data.db.ReservationDaoListenerImpl
+import woowacourse.movie.data.db.ReservationDatabase
 import woowacourse.movie.databinding.FragmentReservationSeatBinding
 import woowacourse.movie.presentation.common.base.BaseFragment
 import woowacourse.movie.presentation.common.custom.CustomAlertDialog
